izmena

+++ app/src/main/java/com/example/marija/mosisproj/Korisnik.java
@@ -13,8 +13,8 @@ class Korisnik {
     public String score;
     public List<String> friends;
     public String picture;
-    public float latitude;
-    public float longitude;
+    public double latitude;
+    public double longitude;
     public String email;
 
     public String getFirstname() {
@@ -68,11 +68,11 @@ class Korisnik {
     public void setPicture(String picture) {
         this.picture = picture;
     }
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
@@ -80,11 +80,11 @@ class Korisnik {
         return email;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
